requirements.txt
pyTelegramBotAPI
requests
python-dotenv
PyYAML